Add files via upload
This is the eia forecast data used by Danish Alvi in the Alvi (2018) paper to demonstrate an application of Probabilistic Graphical Models
in forecasting Crude Oil Price. This dataset is used to compare the model's predictions against the crude oil price and eia's forecast values for the crude oil price as well. This data ranges from 01/10/2015 up till 01/01/2018.